requirements.txt
aiogram==3.10.0
gspread==6.0.0
google-auth==2.25.2
openai==1.12.0
python-dotenv==1.0.0
aiohttp==3.9.1